New Module For Organization and More
Added a new module web_scraping.py for organization. This is going to
become the new web_scraper.py while that will become a main of sorts.

- Moved old function find_table_course

In addition to this change, the ability to get information regarding a
course was also added in this module. This new OOP is necessary for
futureproofing and increasing the ability to play around with the code.

- Added create_course
- Added function to create a dictionary which helps redirect to right
  page for course information