Finalized Course and Program Classes
This will likely be the final update regarding these 2 classes. They
have all the methods they need to hold all relevant info.

- Added _type attribute to Course
- Changed get_course_type to get _type
This was previously returning the subject which the course represented
rather than what type of class it was. This has been changed.

- Added get_course_subject to Course
This is clearer for use within the rest of the project