Added Course Type Finding and Fixed Name Finding
Previously, there was a bug which was a result of bad parsing. It parsed
the name as a culmination of a courses actual name and course type. This
has now been changed so both pieces of information are seperate.
Similarly, now course_type is passed into the Course constructer in
create_course.