Version 1.0 Out Now!
This has been a long time coming. Now there is an interactive version of
the code! Major changes include, but are not limited to:
- Gui which allows real time interaction and Program/Courses addition!
- Gui which you can see a course's description through!

Future updates will be focused on the following:
- Ability to sort courses via tags
- Ability to highlight courses of a certain program

There are lots of bugs still within this version of the code but it's a
good starting point. Have fun out there!