Restructured GUI with new Applet Class
Previously, there was a mainframe class which was taking some of the
frames and storing them into a class. This has been altered to the new
and improved Applet class which has more information and all the update
methods which allows for less parameterization and better usability
across modules.
- Added Applet Class

To work with this change, web_scraping and run have also changed
- Changed imports on web_scraping.py and run.py
- Update functions are now called as methods on an Applet objects and
  changed in web_scraping.py
- Set functions are now called as methods on Applet objects and removed
  from run.py