JSON Serialization and path changes
First and foremost, JSON serializaiton and deserialization was added to
the program as it is a consistent data structure which is usable. This
resulted in the following changes.
- Data/generating_scripts/course_generator now outputs JSON object to
  subject_to_dep_num.txt
- source/web_scraping create_subject_to_department_number only
  deserializes a JSON object now.

A test for this has been added in a new file called test_data. This
brought out some more changes as listed below.
- Imports being changed so they are no longer relative